requirements.txt
flask==3.0.2
openai==1.12.0
python-dotenv==1.0.1
graphviz==0.20.1 